Update PriceRepository: Add Creation

+++ Core/Model/Price.cs
@@ -5,8 +5,8 @@ namespace OnlineShop.Core.Model
     public record Price
     {
         public long Id { get; set; }
-        public decimal Value { get; set; }
-        public decimal Discount { get; set; }
+        public float Value { get; set; }
+        public float Discount { get; set; }
         public DateTime SettingDate { get; set; }
     }
 }
